+++ rechu/command/__init__.py
@@ -3,7 +3,9 @@ Subcommand collection package.
 """
 
 from .base import Base
+from .alembic import Alembic
 from .create import Create
+from .delete import Delete
 from .new import New
 from .read import Read