+++ rechu/command/create.py
@@ -4,7 +4,6 @@ Database schema creation subcommand.
 
 from .base import Base
 from ..database import Database
-from ..models.base import Base as ModelBase
 
 @Base.register("create")
 class Create(Base):
@@ -12,6 +11,11 @@ class Create(Base):
     Create the database with the database schema.
     """
 
+    subparser_keywords = {
+        'help': 'Create the database and schema',
+        'description': 'Create database schema tables at the configured URI.'
+    }
+
     def run(self) -> None:
-        with Database() as session:
-            ModelBase.metadata.create_all(session.get_bind())
+        database = Database()
+        database.create_schema()