+++ rechu/models/receipt.py
@@ -19,17 +19,20 @@ class Receipt(Base): # pylint: disable=too-few-public-methods
     filename: Mapped[str] = mapped_column(String(255), primary_key=True)
     updated: Mapped[datetime.datetime]
     date: Mapped[datetime.date]
-    shop: Mapped[str] = mapped_column(ForeignKey("shop.key"))
+    shop: Mapped[str] = mapped_column(String(32)) # shop.key
     products: Mapped[list["ProductItem"]] = \
-        relationship(back_populates="receipt", cascade="all, delete-orphan")
+        relationship(back_populates="receipt", cascade="all, delete-orphan",
+                     passive_deletes=True)
     discounts: Mapped[list["Discount"]] = \
-        relationship(cascade="all, delete-orphan")
+        relationship(cascade="all, delete-orphan", passive_deletes=True)
 
 # Association table for products involved in discounts
 DiscountItems = Table("receipt_discount_products", Base.metadata,
-                      Column("discount_id", ForeignKey('receipt_discount.id'),
+                      Column("discount_id", ForeignKey('receipt_discount.id',
+                                                       ondelete='CASCADE'),
                              primary_key=True),
-                      Column("product_id", ForeignKey('receipt_product.id'),
+                      Column("product_id", ForeignKey('receipt_product.id',
+                                                      ondelete='CASCADE'),
                              primary_key=True))
 
 class ProductItem(Base): # pylint: disable=too-few-public-methods
@@ -40,7 +43,8 @@ class ProductItem(Base): # pylint: disable=too-few-public-methods
     __tablename__ = "receipt_product"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    receipt_key: Mapped[str] = mapped_column(ForeignKey('receipt.filename'))
+    receipt_key: Mapped[str] = mapped_column(ForeignKey('receipt.filename',
+                                                        ondelete='CASCADE'))
     receipt: Mapped[Receipt] = relationship(back_populates="products")
 
     quantity: Mapped[str]
@@ -48,7 +52,8 @@ class ProductItem(Base): # pylint: disable=too-few-public-methods
     price: Mapped[Price]
     discount_indicator: Mapped[Optional[str]]
     discounts: Mapped[list["Discount"]] = \
-        relationship(secondary=DiscountItems, back_populates="items")
+        relationship(secondary=DiscountItems, back_populates="items",
+                     passive_deletes=True)
 
 class Discount(Base): # pylint: disable=too-few-public-methods
     """
@@ -58,10 +63,12 @@ class Discount(Base): # pylint: disable=too-few-public-methods
     __tablename__ = "receipt_discount"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    receipt_key: Mapped[str] = mapped_column(ForeignKey('receipt.filename'))
+    receipt_key: Mapped[str] = mapped_column(ForeignKey('receipt.filename',
+                                                        ondelete='CASCADE'))
     receipt: Mapped[Receipt] = relationship(back_populates="discounts")
 
     label: Mapped[str]
     price_decrease: Mapped[Price]
     items: Mapped[list[ProductItem]] = \
-        relationship(secondary=DiscountItems, back_populates="discounts")
+        relationship(secondary=DiscountItems, back_populates="discounts",
+                     passive_deletes=True)